- refactoring function getVersionGL() - added mode with without flut window for non WinOS (macros "__GLUT_WINDOW__")

+++ Daniel2.SimplePlayerGL/GPURenderGL.h
@@ -37,7 +37,7 @@ private:
 private:
 	BOOL CreateGL();
 	
-	float getVersionGL();
+	double getVersionGL();
 
 	int gpu_InitGLBuffers();
 	int gpu_DestroyGLBuffers();
